Creature events.
