+++ NWN.Anvil/src/main/API/Events/Game/CreatureEvents/CreatureEvents.OnDeath.cs
@@ -1,0 +1,47 @@
+using System;
+using Anvil.API.Events;
+using NWN.Core;
+
+namespace Anvil.API.Events
+{
+  /// <summary>
+  /// Events for <see cref="NwCreature"/>.
+  /// </summary>
+  public static partial class CreatureEvents
+  {
+    /// <summary>
+    /// Triggered by <see cref="NwCreature"/> when killed by <see cref="NwGameObject"/>.
+    /// </summary>
+    [GameEvent(EventScriptType.CreatureOnDeath)]
+    public sealed class OnDeath : IEvent
+    {
+      /// <summary>
+      /// Gets the <see cref="NwCreature"/> that is killed.
+      /// </summary>
+      public NwCreature KilledCreature { get; } = NWScript.OBJECT_SELF.ToNwObject<NwCreature>();
+
+      /// <summary>
+      /// Gets the <see cref="NwGameObject"/> that killed <see cref="NwCreature"/>.
+      /// </summary>
+      public NwGameObject Killer { get; } = NWScript.GetLastKiller().ToNwObject<NwGameObject>();
+
+      NwObject IEvent.Context
+      {
+        get => KilledCreature;
+      }
+    }
+  }
+}
+
+namespace Anvil.API
+{
+  public sealed partial class NwCreature
+  {
+    /// <inheritdoc cref="CreatureEvents.OnDeath"/>
+    public event Action<CreatureEvents.OnDeath> OnDeath
+    {
+      add => EventService.Subscribe<CreatureEvents.OnDeath, GameEventFactory, GameEventFactory.RegistrationData>(this, new GameEventFactory.RegistrationData(this), value);
+      remove => EventService.Unsubscribe<CreatureEvents.OnDeath, GameEventFactory>(this, value);
+    }
+  }
+}
