+++ NWN.Anvil/src/main/API/Events/Game/CreatureEvents/CreatureEvents.OnSpawn.cs
@@ -1,0 +1,42 @@
+using System;
+using Anvil.API.Events;
+using NWN.Core;
+
+namespace Anvil.API.Events
+{
+  /// <summary>
+  /// Events for <see cref="NwCreature"/>.
+  /// </summary>
+  public static partial class CreatureEvents
+  {
+    /// <summary>
+    /// Triggered by <see cref="NwCreature"/> upon spawning into the game.
+    /// </summary>
+    [GameEvent(EventScriptType.CreatureOnSpawnIn)]
+    public sealed class OnSpawn : IEvent
+    {
+      /// <summary>
+      /// Gets the <see cref="NwCreature"/> that has spawned into the game.
+      /// </summary>
+      public NwCreature Creature { get; } = NWScript.OBJECT_SELF.ToNwObject<NwCreature>();
+
+      NwObject IEvent.Context
+      {
+        get => Creature;
+      }
+    }
+  }
+}
+
+namespace Anvil.API
+{
+  public sealed partial class NwCreature
+  {
+    /// <inheritdoc cref="CreatureEvents.OnSpawn"/>
+    public event Action<CreatureEvents.OnSpawn> OnSpawn
+    {
+      add => EventService.Subscribe<CreatureEvents.OnSpawn, GameEventFactory, GameEventFactory.RegistrationData>(this, new GameEventFactory.RegistrationData(this), value);
+      remove => EventService.Unsubscribe<CreatureEvents.OnSpawn, GameEventFactory>(this, value);
+    }
+  }
+}
