+++ NWN.Anvil/src/main/API/Events/Game/CreatureEvents/CreatureEvents.OnSpellCastAt.cs
@@ -1,0 +1,63 @@
+using System;
+using Anvil.API.Events;
+using NWN.Core;
+
+namespace Anvil.API.Events
+{
+  /// <summary>
+  /// Events for <see cref="NwCreature"/>.
+  /// </summary>
+  public static partial class CreatureEvents
+  {
+    /// <summary>
+    /// Triggered by <see cref="NwCreature"/> when a spell is cast upon it.
+    /// </summary>
+    [GameEvent(EventScriptType.CreatureOnSpellCastAt)]
+    public sealed class OnSpellCastAt : IEvent
+    {
+      /// <summary>
+      /// Gets the <see cref="NwCreature"/> targeted by this spell.
+      /// </summary>
+      public NwCreature Creature { get; } = NWScript.OBJECT_SELF.ToNwObject<NwCreature>();
+
+      /// <summary>
+      /// Gets the <see cref="NwGameObject"/> of this spell. Returns null from an area of effect.
+      /// </summary>
+      public NwGameObject Caster { get; } = NWScript.GetLastSpellCaster().ToNwObject<NwGameObject>();
+
+      /// <summary>
+      /// Gets the <see cref="Spell"/>  that was cast.
+      /// </summary>
+      public Spell Spell { get; } = (Spell)NWScript.GetLastSpell();
+
+      /// <summary>
+      /// Gets a value indicating whether this spell is considered harmful.
+      /// </summary>
+      public bool Harmful { get; } = NWScript.GetLastSpellHarmful().ToBool();
+
+      NwObject IEvent.Context
+      {
+        get => Creature;
+      }
+
+      public static void Signal(NwObject caster, NwCreature target, Spell spell, bool harmful = true)
+      {
+        Event nwEvent = NWScript.EventSpellCastAt(caster, (int)spell, harmful.ToInt());
+        NWScript.SignalEvent(target, nwEvent);
+      }
+    }
+  }
+}
+
+namespace Anvil.API
+{
+  public sealed partial class NwCreature
+  {
+    /// <inheritdoc cref="CreatureEvents.OnSpellCastAt"/>
+    public event Action<CreatureEvents.OnSpellCastAt> OnSpellCastAt
+    {
+      add => EventService.Subscribe<CreatureEvents.OnSpellCastAt, GameEventFactory, GameEventFactory.RegistrationData>(this, new GameEventFactory.RegistrationData(this), value);
+      remove => EventService.Unsubscribe<CreatureEvents.OnSpellCastAt, GameEventFactory>(this, value);
+    }
+  }
+}
