+++ NWN.Anvil/src/main/API/Events/Game/CreatureEvents/CreatureEvents.OnUserDefined.cs
@@ -1,0 +1,44 @@
+using System;
+using Anvil.API.Events;
+using NWN.Core;
+
+namespace Anvil.API.Events
+{
+  /// <summary>
+  /// Events for <see cref="NwCreature"/>.
+  /// </summary>
+  public static partial class CreatureEvents
+  {
+    [GameEvent(EventScriptType.CreatureOnUserDefinedEvent)]
+    public sealed class OnUserDefined : IEvent
+    {
+      public int EventNumber { get; } = NWScript.GetUserDefinedEventNumber();
+
+      public NwCreature Creature { get; } = NWScript.OBJECT_SELF.ToNwObject<NwCreature>();
+
+      NwObject IEvent.Context
+      {
+        get => Creature;
+      }
+
+      public static void Signal(NwCreature creature, int eventId)
+      {
+        Event nwEvent = NWScript.EventUserDefined(eventId);
+        NWScript.SignalEvent(creature, nwEvent);
+      }
+    }
+  }
+}
+
+namespace Anvil.API
+{
+  public sealed partial class NwCreature
+  {
+    /// <inheritdoc cref="CreatureEvents.OnUserDefined"/>
+    public event Action<CreatureEvents.OnUserDefined> OnUserDefined
+    {
+      add => EventService.Subscribe<CreatureEvents.OnUserDefined, GameEventFactory, GameEventFactory.RegistrationData>(this, new GameEventFactory.RegistrationData(this), value);
+      remove => EventService.Unsubscribe<CreatureEvents.OnUserDefined, GameEventFactory>(this, value);
+    }
+  }
+}
